requirements.txt
annotated-types==0.7.0
anyio==4.7.0
attrs==24.3.0
cachetools==5.5.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
fastapi==0.115.6
google-api-core==2.24.0
google-auth==2.37.0
google-cloud-vision==3.9.0
googleapis-common-protos==1.66.0
grpcio==1.68.1
grpcio-status==1.68.1
h11==0.14.0
idna==3.10
imageio==2.36.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
lazy_loader==0.4
llvmlite==0.43.0
networkx==3.4.2
numba==0.60.0
numpy==2.0.2
opencv-python-headless==4.10.0.84
packaging==24.2
pillow==11.1.0
platformdirs==4.3.6
pooch==1.8.2
proto-plus==1.25.0
protobuf==5.29.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.4
pydantic_core==2.27.2
PyMatting==1.1.13
python-dotenv==1.0.1
referencing==0.35.1
rembg==2.0.61
requests==2.32.3
rpds-py==0.22.3
rsa==4.9
scikit-image==0.25.0
scipy==1.15.0
sniffio==1.3.1
starlette==0.41.3
tifffile==2024.12.12
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
onnxruntime
python-multipart
